Renamed files for better docs and added README
