+++ README.md
@@ -1,0 +1,15 @@
+# Future Features
+
+## TODO's 
+
+### Add a drop down to the nav bar for the attendance feature.
+- Drop down will be titled Attendance or something similiar 
+- Drop down will contain a list of current herds
+- User will be able to select which herd they would like to take attendance for
+- Once herd is selected a new page will be opened where attendance of the herd will be taken
+
+### Cattle attendance page
+- User will mark off accounted for cattle with a check box
+- At the bottom a confirm button will be availabe 
+- Once confirmed is clicked the attendance will be saved and be able to be viewed at a later date
+- User will be redirected to the page for the herd they just took attendace for
